requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.3
google-api-core==2.11.0
google-auth==2.16.2
google-cloud-bigquery==3.7.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
packaging==23.0
pandas==1.5.3
plotly==5.13.1
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
rsa==4.9
six==1.16.0
tenacity==8.2.2
urllib3==1.26.15
Werkzeug==2.2.3
